follow automake version 2.0 update advising
从automake更新到1.14.1开始，需要在autoconf的配置中
加上 subdir-objects 选项才能顺利编译。
不然 autoreconf --install 会报错。

From the automake-1.14.1 NEWS file:

The next major Automake version (2.0) will unconditionally activate
the 'subdir-objects' option. In order to smooth out the transition,
we now give a warning (in the category 'unsupported') whenever a
source file is present in a subdirectory but the 'subdir-object' is
not enabled. For example, the following usage will trigger such a
warning:
    bin_PROGRAMS = sub/foo
    sub_foo_SOURCES = sub/main.c sub/bar.c


ref：https://github.com/liuqun/njit8021xclient/issues/13
ref: http://stackoverflow.com/questions/21609580/autotools-build-fails-due-to-subdir-objects-option-in-am-init-automake (所以我觉得修改 Makefile.am 兼容性更好一些)